fix(db): use BigInteger for monitoring byte columns (PostgreSQL overflow)
Integer columns overflow at ~2.1GB on PostgreSQL. Switch memory/network
byte columns to BigInteger and add migration for existing databases.

+++ backend/alembic/versions/004_add_monitoring_columns.py
@@ -28,7 +28,7 @@ def upgrade() -> None:
 
     # Add new column to memory_samples
     with op.batch_alter_table('memory_samples', schema=None) as batch_op:
-        batch_op.add_column(sa.Column('baluhost_memory_bytes', sa.Integer(), nullable=True))
+        batch_op.add_column(sa.Column('baluhost_memory_bytes', sa.BigInteger(), nullable=True))
 
 
 def downgrade() -> None:
